requirements.txt
flet>=0.25.0
youtube-transcript-api>=0.6.0
google-generativeai>=0.8.0
